website/requirements.txt
easygoogletranslate==0.0.4
Flask==2.2.5
flask_mail==0.9.1
flask_oauthlib==0.9.6
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.2
pandas==2.1.4
plotly==5.18.0
prophet==1.1.5
python-dotenv==1.0.0
Requests==2.31.0
selenium==4.16.0
Werkzeug==2.2.2